% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryriskindex-package.R
\docType{package}
\name{countryriskindex-package}
\alias{countryriskindex}
\alias{countryriskindex-package}
\title{countryriskindex: Country Health Risk Calculation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

countryriskindex is an R package for computing the Country Health Risk Index (CHRI), a composite indicator designed to assess national health risks by integrating multiple dimensions of public health vulnerability.
}
\author{
\strong{Maintainer}: Andrea Pedot \email{andrea.pedot@unimi.it} (\href{https://orcid.org/0000-0001-7050-1378}{ORCID})

}
\keyword{internal}
