% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{create_lollipop_plot_comp}
\alias{create_lollipop_plot_comp}
\title{Lollipop Plot Comparing Risk Score with CRPI Index}
\usage{
create_lollipop_plot_comp(
  risk_score,
  crpi,
  x_col = "overall_risk",
  y_col = "Index",
  sort_by_difference = FALSE,
  color_x = "lightblue",
  color_y = "lightpink"
)
}
\arguments{
\item{risk_score}{A data frame with ISO_A3, country, and a risk score column.}

\item{crpi}{A data frame with iso_a3 and CRPI index columns.}

\item{x_col}{Name of the risk score column in \code{risk_score}.}

\item{y_col}{Name of the CRPI index column in \code{crpi}.}

\item{sort_by_difference}{Logical. Whether to sort by absolute difference.}

\item{color_x}{Color for the x_col points.}

\item{color_y}{Color for the y_col points.}
}
\value{
A ggplot2 lollipop plot.
}
\description{
Joins a risk score dataset with a CRPI dataset and creates a lollipop plot
comparing the specified risk score column with the CRPI index.
}
