% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{create_lollipop_plot_shift}
\alias{create_lollipop_plot_shift}
\title{Lollipop Plot Comparing Overall Risk and Severity-Adjusted Risk}
\usage{
create_lollipop_plot_shift(
  risk_score,
  sort_by_difference = FALSE,
  color_x = "lightblue",
  color_y = "deepskyblue"
)
}
\arguments{
\item{risk_score}{A data frame with columns ISO_A3, country, overall_risk,
and severity_adjusted_risk.}

\item{sort_by_difference}{Logical. Whether to sort by difference.}

\item{color_x}{Color for overall_risk points.}

\item{color_y}{Color for severity-adjusted risk points.}
}
\value{
A ggplot2 lollipop plot.
}
\description{
Creates a lollipop plot comparing overall risk and severity-adjusted risk
scores
for each country.
}
