% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{create_outbreak_concern}
\alias{create_outbreak_concern}
\title{Plot a single stacked bar chart for outbreak concern questions}
\usage{
create_outbreak_concern(
  long,
  concern_levels = c("Very unconcerned", "Somewhat unconcerned",
    "Neither concerned nor unconcerned", "Somewhat concerned", "Very concerned"),
  outbreak_labels = c("Food/Waterborne Disease Concern", "Food/Waterborne Outbreak Risk",
    "STDs Outbreak Concern", "STDs Transmission Risk",
    "Hemorrhagic Fever Outbreak Concern", "Hemorrhagic Fever Risk", "Zoonoses Concern",
    "Vector-borne Disease Outbreak", "Vector-borne Disease Risk",
    "Airborne Disease Outbreak", "Airborne Disease Risk")
)
}
\arguments{
\item{long}{The full long-format tibble from \code{pivot_disease_data()}.}

\item{concern_levels}{Ordered vector of concern levels.}

\item{outbreak_labels}{Vector of cleaned outbreak question labels (length = 11).}
}
\value{
A ggplot object.
}
\description{
Visualizes concern levels across all outbreak-related questions in one plot.
Each question is a separate bar, filled by proportion of concern levels.
}
