% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{create_pca_biplot}
\alias{create_pca_biplot}
\title{Biplot of countries in the first two principal components}
\usage{
create_pca_biplot(pca_result, colour_by = NULL, label_n = 8)
}
\arguments{
\item{pca_result}{Output of \code{\link[=run_pca_validation]{run_pca_validation()}}.}

\item{colour_by}{Optional column in \code{pca_result$scores} used to colour points.}

\item{label_n}{Number of most extreme countries to label on each axis.}
}
\value{
A ggplot2 object.
}
\description{
Biplot of countries in the first two principal components
}
