% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{create_pca_loadings}
\alias{create_pca_loadings}
\title{Loadings heatmap for the leading principal components}
\usage{
create_pca_loadings(pca_result, n_pc = 4)
}
\arguments{
\item{pca_result}{Output of \code{\link[=run_pca_validation]{run_pca_validation()}}.}

\item{n_pc}{Number of components to display.}
}
\value{
A ggplot2 object.
}
\description{
Loadings heatmap for the leading principal components
}
