% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_.R
\name{create_pca_scree}
\alias{create_pca_scree}
\title{Scree plot of explained variance}
\usage{
create_pca_scree(pca_result, max_pc = 10)
}
\arguments{
\item{pca_result}{Output of \code{\link[=run_pca_validation]{run_pca_validation()}}.}

\item{max_pc}{Maximum number of components to display.}
}
\value{
A ggplot2 object.
}
\description{
Scree plot of explained variance
}
