% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_cap_bars}
\alias{createimg_cap_bars}
\title{Lack of Coping Capacity plot}
\usage{
createimg_cap_bars(risk_score = NULL)
}
\arguments{
\item{risk_score}{Data with capacity values}
}
\value{
A ggplot object
}
\description{
Lack of Coping Capacity plot
}
