% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_daly_bars}
\alias{createimg_daly_bars}
\title{DALY Rate plot}
\usage{
createimg_daly_bars(gbd_rates = NULL)
}
\arguments{
\item{gbd_rates}{Data with DALY values}
}
\value{
A ggplot object
}
\description{
DALY Rate plot
}
