% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_risk_bars}
\alias{createimg_risk_bars}
\title{Risk Score plot}
\usage{
createimg_risk_bars(risk_score = NULL)
}
\arguments{
\item{risk_score}{Data with risk values}
}
\value{
A ggplot object
}
\description{
Risk Score plot
}
