% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_risk_hist}
\alias{createimg_risk_hist}
\title{Create a Histogram of Health Risk Scores}
\usage{
createimg_risk_hist(risk_score = NULL)
}
\arguments{
\item{risk_score}{A data frame containing a column named 'overall_risk' and
optionally 'risk'.}
}
\value{
A ggplot2 histogram object.
}
\description{
Generates a histogram of the overall health risk scores using ggplot2.
The histogram is colored by the 'risk' variable and hides the legend.
}
