% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_u5mr_bars}
\alias{createimg_u5mr_bars}
\title{Under-5 Mortality Rate plot}
\usage{
createimg_u5mr_bars(u5_mr = NULL)
}
\arguments{
\item{u5_mr}{Data with U5MR values}
}
\value{
A ggplot object
}
\description{
Under-5 Mortality Rate plot
}
