% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createimg_vul_bars}
\alias{createimg_vul_bars}
\title{Vulnerability plot}
\usage{
createimg_vul_bars(risk_score = NULL)
}
\arguments{
\item{risk_score}{Data with vulnerability values}
}
\value{
A ggplot object
}
\description{
Vulnerability plot
}
