% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_DALY_avg}
\alias{createmap_DALY_avg}
\title{Interactive Map of DALY Rate}
\usage{
createmap_DALY_avg(gbd_rates = NULL)
}
\arguments{
\item{gbd_rates}{A data frame with an \code{"All causes"} column and ISO_A3-coded
countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of DALY rates for all causes using a blue color
scale.
}
