% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_DALY_score}
\alias{createmap_DALY_score}
\title{Interactive Map of DALY Quantiles}
\usage{
createmap_DALY_score(gbd_rates = NULL)
}
\arguments{
\item{gbd_rates}{A data frame with an \code{"All causes"} column and ISO_A3-coded
countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of DALY quantile scores using a plasma color scale.
}
