% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_severity_score}
\alias{createmap_severity_score}
\title{Interactive Map of Crisis Severity Index}
\usage{
createmap_severity_score(db = NULL)
}
\arguments{
\item{db}{A data frame with a \code{"severity_index"} column and ISO_A3-coded
countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of INFORM Severity Index scores using an accent
color scale.
}
