% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_u5mr_avg}
\alias{createmap_u5mr_avg}
\title{Interactive Map of Under-5 Mortality Rate}
\usage{
createmap_u5mr_avg(u5_mr = NULL)
}
\arguments{
\item{u5_mr}{A data frame with a \code{"Total"} column and ISO_A3-coded countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of average under-5 mortality rates by country using
a yellow-red color scale.
}
