% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_u5mr_did}
\alias{createmap_u5mr_did}
\title{Interactive Map of U5MR Intra-country Differences}
\usage{
createmap_u5mr_did(u5_mr = NULL)
}
\arguments{
\item{u5_mr}{A data frame with a \code{"did"} column and ISO_A3-coded countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of intra-country differences in under-5 mortality
rates using a viridis color scale.
}
