% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{createmap_vul_score}
\alias{createmap_vul_score}
\title{Interactive Map of Vulnerability Score}
\usage{
createmap_vul_score(risk_score = NULL)
}
\arguments{
\item{risk_score}{A data frame with a \code{"vulnerability_score"} column and
ISO3-coded countries.}
}
\value{
A \code{leaflet} map object.
}
\description{
Displays a choropleth map of normalized vulnerability scores using a viridis
color scale.
}
