% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_.R
\name{decomposition_labels}
\alias{decomposition_labels}
\title{Human-readable labels for decomposition items}
\usage{
decomposition_labels()
}
\value{
A named character vector.
}
\description{
Human-readable labels for decomposition items
}
