% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_.R
\name{import_crpi_excel}
\alias{import_crpi_excel}
\title{Import Excel from previous risk assessment}
\usage{
import_crpi_excel(path)
}
\arguments{
\item{path}{Path to the Excel file.}
}
\value{
A tibble with the imported data.
}
\description{
Import Excel from previous risk assessment
}
