% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{invert_0_10}
\alias{invert_0_10}
\title{Invert a 0–10 scale}
\usage{
invert_0_10(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Inverted numeric vector.
}
\description{
Converts values on a 0–10 scale to their inverse (10 - x).
}
