% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{normalise_quantiles}
\alias{normalise_quantiles}
\title{Normalize values between the 1st and 99th percentiles}
\usage{
normalise_quantiles(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A normalized numeric vector.
}
\description{
Scales a numeric vector so that the 1st percentile becomes 0 and the 99th becomes 1.
}
