% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{plot_all_diseases_from_long}
\alias{plot_all_diseases_from_long}
\title{Plot all diseases using long-format data}
\usage{
plot_all_diseases_from_long(
  long,
  concern_levels = c("Very unconcerned", "Somewhat unconcerned",
    "Neither concerned nor unconcerned", "Somewhat concerned", "Very concerned")
)
}
\arguments{
\item{long}{The full long-format tibble from \code{pivot_disease_data()}.}

\item{concern_levels}{Ordered vector of concern levels.}
}
\value{
A named list of ggplot objects.
}
\description{
Plot all diseases using long-format data
}
