% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_.R
\name{plot_all_questions_from_long}
\alias{plot_all_questions_from_long}
\title{Plot all 10 questions across diseases using combined long-format data}
\usage{
plot_all_questions_from_long(
  long,
  concern_levels = c("Very unconcerned", "Somewhat unconcerned",
    "Neither concerned nor unconcerned", "Somewhat concerned", "Very concerned"),
  question_labels = c("15m diagnosis delay", "1h diagnosis delay", "4h diagnosis delay",
    "8h diagnosis delay", "1h transport delay", "4h transport delay",
    "8h transport delay", "1 day transport delay", "Patient life impact",
    "Patient wellbeing impact")
)
}
\arguments{
\item{long}{A combined tibble from multiple diseases (output of \code{pivot_disease_data()}),
with an added \code{disease} column.}

\item{concern_levels}{Ordered vector of concern levels.}

\item{question_labels}{Vector of question titles (length = 10).}
}
\value{
A named list of ggplot objects.
}
\description{
Plot all 10 questions across diseases using combined long-format data
}
