% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_.R
\name{preprocess_gbd_rates_by_cause}
\alias{preprocess_gbd_rates_by_cause}
\title{Preprocess GBD DALY Rates by Cause}
\usage{
preprocess_gbd_rates_by_cause(db = NULL, loc_keys = NULL)
}
\arguments{
\item{db}{A data frame with GBD DALY data including \code{location_id},
\code{cause_id}, and \code{val}.}

\item{loc_keys}{A data frame mapping \verb{Location ID} to \code{ISO_A3} codes.}
}
\value{
A data frame with DALY rates by cause for each country.
}
\description{
Processes Global Burden of Disease (GBD) DALY data to extract
disease-specific rates.
The function:
\itemize{
\item Maps GBD location IDs to ISO A3 codes.
\item Filters for selected cause IDs and reshapes the data to wide format.
\item Computes a composite "Other NCDs" category by summing DALYs for a
predefined set of causes.
\item Merges the selected causes and the composite category into a single
data frame.
}
}
