% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_.R
\name{preprocess_inform}
\alias{preprocess_inform}
\title{Preprocess INFORM Risk and Coping Capacity Data}
\usage{
preprocess_inform(db_risk = NULL, db_lcc = NULL)
}
\arguments{
\item{db_risk}{A data frame with INFORM Risk data.}

\item{db_lcc}{A data frame with Lack of Coping Capacity data.}
}
\value{
A merged data frame with risk, vulnerability, and infrastructure
indicators.
}
\description{
Merges INFORM Risk and Lack of Coping Capacity (LCC) datasets to extract and
transform
indicators relevant to vulnerability assessment. The function:
\itemize{
\item Extracts numeric indicators from the INFORM Risk dataset.
\item Removes metadata rows from the LCC dataset and converts key
indicators to numeric.
\item Joins both datasets by ISO3 country code.
}
}
