% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_.R
\name{risk_components}
\alias{risk_components}
\title{Components of the index, in reporting order}
\usage{
risk_components()
}
\value{
A character vector.
}
\description{
Components of the index, in reporting order
}
