% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Country Risk Explorer Shiny Application}
\usage{
run_app(raw_data, radar_data, corrected_radar_data)
}
\arguments{
\item{raw_data}{A data frame containing the full dataset with country-level
health and risk indicators. This dataset is used across all visualizations.}

\item{radar_data}{A data frame containing a subset of the raw data
specifically
formatted for radar chart visualization. It should include numeric indicators
relevant to health risk dimensions.}

\item{corrected_radar_data}{A data frame similar in structure to \code{radar_data},
but with severity-adjusted values used to compare original and corrected
risk profiles.}
}
\value{
This function does not return a value. It launches a Shiny
application in the user's default web browser.
}
\description{
This function initializes and launches an interactive Shiny web application
designed to explore and visualize country-level health and crisis risk data.
The app includes multiple interactive visualizations such as a lollipop plot,
3D scatterplot, radar chart, risk distribution histogram, and an adjusted risk
map.
}
