% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_.R
\name{snapshot_columns}
\alias{snapshot_columns}
\title{Columns retained in a run snapshot}
\usage{
snapshot_columns()
}
\value{
A character vector of column names.
}
\description{
Columns retained in a run snapshot
}
