% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{add_hazard_score}
\alias{add_hazard_score}
\title{Compute Hazard Score from GBD Rates}
\usage{
add_hazard_score(df)
}
\arguments{
\item{df}{A data frame containing a column named \code{"All causes"}.}
}
\value{
The input data frame with an added column \code{hazard_score}.
}
\description{
Calculates a normalized hazard score based on the "All causes" DALY rate from
GBD data.
The score is computed by dividing each value by the maximum observed value.
}
