% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Compute the geometric mean}
\usage{
geometric_mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Geometric mean as a single numeric value.
}
\description{
Calculates the geometric mean of a numeric vector, ignoring NAs.
}
