% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utils.R, R/utils.R
\name{invert_0_100}
\alias{invert_0_100}
\title{Invert a 0–100 scale}
\usage{
invert_0_100(x)

invert_0_100(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Inverted numeric vector.

Inverted numeric vector.
}
\description{
Converts values on a 0–100 scale to their inverse (100 - x).

Converts values on a 0-100 scale to their inverse (100 - x).
}
